+++ CHANGES.rst
@@ -9,6 +9,7 @@ Added
 - Configuration option to disable the forced usage of OIDC nonce :pr:`143`
 - Validate phone numbers with a regex :pr:`146`
 - Email verification :issue:`41` :pr:`147`
+- Account registration :issue:`55` :pr:`133` :pr:`148`
 
 Fixed
 *****